+++ src/test/java/com/theoryinpractise/coffeescript/RhinoCompilerTest.java
@@ -2,13 +2,15 @@ package com.theoryinpractise.coffeescript;
 
 import com.google.common.base.Function;
 import com.google.common.collect.Iterators;
+import com.theoryinpractise.coffeescript.CoffeeScriptCompilerMojo;
+import com.theoryinpractise.coffeescript.compiler.RhinoCompiler;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
 import javax.annotation.Nullable;
 import java.util.Iterator;
 
-public class CoffeeScriptCompilerTest {
+public class RhinoCompilerTest {
 
     @DataProvider
     public Iterator<Object[]> provideVersions() {
@@ -23,7 +25,7 @@ public class CoffeeScriptCompilerTest {
 
     @Test(dataProvider = "provideVersions")
     public void testCompiler(final String version) {
-        CoffeeScriptCompiler compiler = new CoffeeScriptCompiler(version, true);
+        RhinoCompiler compiler = new RhinoCompiler(version, true);
 
     }
